# openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TeamCloud",
    "description": "API for working with a TeamCloud instance.",
    "contact": {
      "name": "TeamCloud Dev Team",
      "url": "https://github.com/microsoft/TeamCloud/issues/new",
      "email": "Markus.Heiliger@microsoft.com"
    },
    "license": {
      "name": "TeamCloud is licensed under the MIT License",
      "url": "https://github.com/microsoft/TeamCloud/blob/main/LICENSE"
    },
    "version": "v1"
  },
  "paths": {
...
    "/orgs/{organizationId}/projects/{projectId}/components/{id}": {
      "get": {
        "tags": [
          "Components"
        ],
        "summary": "Gets a Project Component.",
        "operationId": "GetComponent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns Project Component",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComponentDataResult"
                }
              }
            }
          },
          "400": {
            "description": "A validation error occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "404": {
            "description": "A Project with the provided projectId was not found, or a Component with the provided id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "delete": {
        "tags": [
          "Components"
        ],
        "summary": "Deletes an existing Project Component.",
        "operationId": "DeleteComponent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Starts deleting the Project Component. Returns a StatusResult object that can be used to track progress of the long-running operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResult"
                }
              }
            }
          },
          "204": {
            "description": "The Project Component was deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComponentDataResult"
                }
              }
            }
          },
          "400": {
            "description": "A validation error occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "404": {
            "description": "A Project with the provided id was not found, or a Component with the provided id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/orgs/{organizationId}/projects/{projectId}/components/{componentId}/tasks": {
      "get": {
        "tags": [
          "ComponentTasks"
        ],
        "summary": "Gets all Component Tasks.",
        "operationId": "GetComponentTasks",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "componentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all Component Tasks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComponentTaskListDataResult"
                }
              }
            }
          },
          "400": {
            "description": "A validation error occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "404": {
            "description": "A Component with the provided componentId was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "ComponentTasks"
        ],
        "summary": "Creates a new Project Component Task.",
        "operationId": "CreateComponentTask",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "componentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComponentTaskDefinition"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created Project Component Task.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComponentTaskDataResult"
                }
              }
            }
          },
          "202": {
            "description": "Starts creating the new Project Component. Returns a StatusResult object that can be used to track progress of the long-running operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResult"
                }
              }
            }
          },
          "400": {
            "description": "A validation error occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "404": {
            "description": "A Project with the provided projectId was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "A Project Component already exists with the id provided in the request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/orgs/{organizationId}/projects/{projectId}/components/{componentId}/tasks/{id}": {
      "get": {
        "tags": [
          "ComponentTasks"
        ],
        "summary": "Gets the Component Task.",
        "operationId": "GetComponentTask",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "componentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a Component Task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComponentTaskDataResult"
                }
              }
            }
          },
          "400": {
            "description": "A validation error occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "404": {
            "description": "A Component Task with the provided id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Component": {
        "required": [
          "creator",
          "id",
          "organization",
          "projectId",
          "slug",
          "templateId",
          "type"
        ],
        "type": "object",
        "properties": {
          "href": {
            "type": "string",
            "nullable": true
          },
          "organization": {
            "type": "string"
          },
          "templateId": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "creator": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "inputJson": {
            "type": "string",
            "nullable": true
          },
          "valueJson": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "enum": [
              "Custom",
              "AzureResource",
              "Environment",
              "GitRepository"
            ],
            "type": "string"
          },
          "resourceId": {
            "type": "string",
            "nullable": true
          },
          "resourceState": {
            "enum": [
              "Pending",
              "Initializing",
              "Provisioning",
              "Succeeded",
              "Failed"
            ],
            "type": "string"
          },
          "deploymentScopeId": {
            "type": "string",
            "nullable": true
          },
          "identityId": {
            "type": "string",
            "nullable": true
          },
          "deleted": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ttl": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "slug": {
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ValidationError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ResultError": {
        "type": "object",
        "properties": {
          "code": {
            "enum": [
              "Unknown",
              "Failed",
              "Conflict",
              "NotFound",
              "ServerError",
              "ValidationError",
              "Unauthorized",
              "Forbidden"
            ],
            "type": "string"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ErrorResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResultError"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ComponentDataResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/Component"
          },
          "location": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "StatusResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "stateMessage": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResultError"
            },
            "nullable": true
          },
          "_trackingId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ComponentTask": {
        "required": [
          "componentId",
          "id",
          "organization",
          "projectId"
        ],
        "type": "object",
        "properties": {
          "organization": {
            "type": "string"
          },
          "componentId": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "requestedBy": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "enum": [
              0,
              1,
              2
            ],
            "type": "integer",
            "format": "int32"
          },
          "typeName": {
            "type": "string",
            "nullable": true
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "started": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "finished": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "inputJson": {
            "type": "string",
            "nullable": true
          },
          "output": {
            "type": "string",
            "nullable": true
          },
          "resourceId": {
            "type": "string",
            "nullable": true
          },
          "resourceState": {
            "enum": [
              "Pending",
              "Initializing",
              "Provisioning",
              "Succeeded",
              "Failed"
            ],
            "type": "string"
          },
          "exitCode": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "id": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ComponentTaskListDataResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ComponentTask"
            },
            "nullable": true,
            "readOnly": true
          },
          "location": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ComponentTaskDefinition": {
        "required": [
          "taskId"
        ],
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string"
          },
          "inputJson": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ComponentTaskDataResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/ComponentTask"
          },
          "location": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://login.microsoftonline.com//oauth2/v2.0/authorize",
            "tokenUrl": "https://login.microsoftonline.com//oauth2/v2.0/token",
            "scopes": {
              "openid": "Sign you in",
              "http://TeamCloud.aztcclitestsix/user_impersonation": "Access the TeamCloud API"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "oauth2": [
        "openid",
        "http://TeamCloud.aztcclitestsix/user_impersonation"
      ]
    }
  ]
}